requirements.txt
sanic~=23.3.0
requests~=2.31.0
python-dotenv~=1.0.0
aioschedule~=0.5.2
aiohttp~=3.8.5